+++ lib/screen/delivery/posting/view/delivery_what_food_page.dart
@@ -1,43 +1,42 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:mukgen_flutter_v1/common/common.dart';
-import 'package:mukgen_flutter_v1/screen/delivery/delivery_what_time_page.dart';
+import 'package:mukgen_flutter_v1/screen/delivery/posting/view/delivery_how_many_page.dart';
 import 'package:mukgen_flutter_v1/widget/mukgen_button.dart';
 import 'package:mukgen_flutter_v1/widget/mukgen_text_field.dart';
 import 'package:transition/transition.dart';
 
-class DeliveryWhereMeetPage extends StatefulWidget {
-  const DeliveryWhereMeetPage({Key? key, required this.menu, required this.participantNumber}) : super(key: key);
+class DeliveryWhatFoodPage extends StatefulWidget {
+  const DeliveryWhatFoodPage({Key? key}) : super(key: key);
 
-  final String menu;
-  final int participantNumber;
   @override
-  State<DeliveryWhereMeetPage> createState() => _DeliveryWhereMeetPageState();
+  State<DeliveryWhatFoodPage> createState() => _DeliveryWhatFoodPageState();
 }
 
-class _DeliveryWhereMeetPageState extends State<DeliveryWhereMeetPage> {
-  late TextEditingController wheremeetController;
+class _DeliveryWhatFoodPageState extends State<DeliveryWhatFoodPage> {
+  late TextEditingController whatfoodController;
 
   bool _isButtonEnabled = false;
 
   @override
   void initState() {
     super.initState();
-    wheremeetController = TextEditingController();
-    wheremeetController.addListener(_updateButtonState);
+    whatfoodController = TextEditingController();
+    whatfoodController.addListener(_updateButtonState);
   }
 
   @override
   void dispose() {
     super.dispose();
-    wheremeetController.dispose();
+    whatfoodController.dispose();
   }
 
   void _updateButtonState() {
     setState(() {
-      _isButtonEnabled = wheremeetController.text.isNotEmpty;
+      _isButtonEnabled = whatfoodController.text.isNotEmpty;
     });
   }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -76,7 +75,7 @@ class _DeliveryWhereMeetPageState extends State<DeliveryWhereMeetPage> {
             padding: EdgeInsets.only(right: 20.0.w),
             alignment: Alignment.bottomRight,
             child: Text(
-              '3 / 4',
+              '1 / 4',
               style: TextStyle(
                 color: MukGenColor.primaryLight2,
                 fontWeight: FontWeight.w600,
@@ -90,7 +89,7 @@ class _DeliveryWhereMeetPageState extends State<DeliveryWhereMeetPage> {
             padding: EdgeInsets.only(left: 20.0.w),
             alignment: Alignment.bottomLeft,
             child: Text(
-              '어디서\n만나실 건가요?',
+              '어떤 음식을\n배달 시키시나요?',
               style: TextStyle(
                 color: MukGenColor.black,
                 fontWeight: FontWeight.w600,
@@ -102,55 +101,34 @@ class _DeliveryWhereMeetPageState extends State<DeliveryWhereMeetPage> {
           SizedBox(height: 40.0.h),
           MukGenTextField(
             width: 353,
-            controller: wheremeetController,
+            controller: whatfoodController,
             fontSize: 20,
             isPwdTextField: false,
             autofocus: true,
             maxLength: null,
-            hintText: "장소",
+            hintText: "음식 이름",
           ),
           const Spacer(),
-          Row(
-            children: [
-              SizedBox(width: 30.0.w),
-              MukGenButton(
-                width: 161.5,
-                height: 55,
-                text: "이전",
-                backgroundColor: MukGenColor.primaryLight3,
-                textColor: MukGenColor.pointBase,
-                fontSize: 16.sp,
-                outlineColor: MukGenColor.pointBase,
-                outlineWidth: 1.0,
-                onPressed: () {
-                  setState(() {
-                    Navigator.of(context).pop();
-                  });
-                },
-              ),
-              SizedBox(width: 10.0.w),
-              MukGenButton(
-                width: 161.5,
-                height: 55,
-                text: "다음",
-                backgroundColor: _isButtonEnabled
-                    ? MukGenColor.pointBase
-                    : MukGenColor.primaryLight2,
-                textColor: MukGenColor.white,
-                fontSize: 16.sp,
-                onPressed: () {
-                  _isButtonEnabled
-                      ? Navigator.push(
-                    context,
-                    Transition(
-                      child: DeliveryWhatTimePage(menu: widget.menu, participantNumber: widget.participantNumber, place: wheremeetController.text,),
-                      transitionEffect: TransitionEffect.RIGHT_TO_LEFT,
-                    ),
-                  )
-                      : null;
-                },
-              ),
-            ],
+          MukGenButton(
+            width: 352,
+            height: 55,
+            text: "다음",
+            backgroundColor: _isButtonEnabled
+                ? MukGenColor.pointBase
+                : MukGenColor.primaryLight2,
+            textColor: MukGenColor.white,
+            fontSize: 16.sp,
+            onPressed: () {
+              _isButtonEnabled
+                  ? Navigator.push(
+                      context,
+                      Transition(
+                        child: DeliveryHowManyPage(menu: whatfoodController.text),
+                        transitionEffect: TransitionEffect.RIGHT_TO_LEFT,
+                      ),
+                    )
+                  : null;
+            },
           ),
           SizedBox(height: 20.0.h),
         ],
